Stack Problem

+++ Edit Distance/Edit Distance/main.cpp
@@ -10,8 +10,8 @@ using namespace std;
 class Solution {
 public:
     int minDistance(string word1, string word2) {
-        int m = word1.size();
-        int n = word2.size();
+        int m = (int) word1.size();
+        int n = (int) word2.size();
         int dp[m + 1][n + 1];
         
         for (int i = 0; i <= m; i++)
